+++ tests/adapter/unit/test_protocol.c
@@ -11,8 +11,9 @@
 
 #define _POSIX_C_SOURCE 200809L
 
-#include "../sdk/adapter/include/cortex_transport.h"
-#include "../sdk/adapter/include/cortex_protocol.h"
+#include "cortex_transport.h"
+#include "cortex_protocol.h"
+#include "cortex_wire.h"
 
 #include <assert.h>
 #include <stdio.h>